Squashed big bug in PAZ.to_obspy() that affected local amplitude calculation.
Was giving different answers for the same channel/input_units depending on whether
the input_units had been transformed or not, because A0 is not automatically
recalculated.  PAZ.to_obspy() now recalculates A0 before assigning "gain"
and "sensitivity" parameters to the PAZ dict that it outputs
Trace.simulate usese PAZ dicts and so this variability gave very different
amplitudes depending on wheter or not I had transformed the input_units.